add changes and new classes

+++ platzi-market-jh/src/main/java/com/platzi/market_jh/domain/Product.java
@@ -1,9 +1,7 @@
 package com.platzi.market_jh.domain;
 
-import com.platzi.market_jh.persistence.entities.Categoria;
-
 public class Product {
-    private int productId;
+    private Long productId;
     private String name;
     private int categoryId;
     private int price;
@@ -11,11 +9,11 @@ public class Product {
     private boolean active;
     private Category category;
 
-    public int getProductId() {
+    public Long getProductId() {
         return productId;
     }
 
-    public void setProductId(int productId) {
+    public void setProductId(Long productId) {
         this.productId = productId;
     }
