+++ platzi-market-jh/src/main/java/com/platzi/market_jh/domain/repository/ProductRepository.java
@@ -8,13 +8,13 @@ import java.util.Optional;
 public interface ProductRepository {
     List<Product> getAll();
 
-    Optional<List<Product>> getByCategory(int categoryId);
+    Optional<List<Product>> getByCategory(Integer categoryId);
 
     Optional<List<Product>> getScarseProducts(int quantity);
 
-    Optional<Product> getProduct(int productId);
+    Optional<Product> getProduct(Long productId);
 
     Product save(Product product);
 
-    void delete(int productId);
+    void delete(Long productId);
 }
